+++ src/index.js
@@ -1,6 +1,5 @@
 const express = require("express");
 const bodyParser = require("body-parser");
-const tasksRepository = require("./tasksRepository");
 
 //cambio nuevo
 const app = express();
@@ -15,7 +14,6 @@ let tasks = [
 
 // Get all tasks
 app.get("/tasks", (req, res) => {
-  const tasks = tasksRepository.getAll();
   res.json(tasks);
 });
 
@@ -26,7 +24,7 @@ app.get("/issue", (req, res) => {
 // Get a specific task
 app.get("/tasks/:id", (req, res) => {
   const taskId = parseInt(req.params.id);
-  const task = tasksRepository((t) => t.id === taskId);
+  const task = tasks.find((t) => t.id === taskId);
 
   if (task) {
     res.json(task);
@@ -38,7 +36,8 @@ app.get("/tasks/:id", (req, res) => {
 // Create a new task
 app.post("/tasks", (req, res) => {
   const newTask = req.body;
-  tasksRepository.createTask(newTask);
+  newTask.id = tasks.length + 1;
+  tasks.push(newTask);
   res.status(201).json(newTask);
 });
 
@@ -46,10 +45,11 @@ app.post("/tasks", (req, res) => {
 app.put("/tasks/:id", (req, res) => {
   const taskId = parseInt(req.params.id);
   const updatedTask = req.body;
-  const task = tasksRepository.updateTask(taskId, updatedTask);
+  const index = tasks.findIndex((t) => t.id === taskId);
 
-  if (task !== -1) {
-    res.json(tasks);
+  if (index !== -1) {
+    tasks[index] = { ...tasks[index], ...updatedTask };
+    res.json(tasks[index]);
   } else {
     res.status(404).json({ error: "Task not found" });
   }
@@ -58,7 +58,7 @@ app.put("/tasks/:id", (req, res) => {
 // Delete a task
 app.delete("/tasks/:id", (req, res) => {
   const taskId = parseInt(req.params.id);
-  tasksRepository.deleteTask(taskId);
+  tasks = tasks.filter((t) => t.id !== taskId);
   res.sendStatus(204);
 });
